+++ android/atsc3_bridge_media_mmt/src/main/java/org/ngbp/libatsc3/middleware/android/mmt/MmtPacketIdContext.java
@@ -81,7 +81,7 @@ public class MmtPacketIdContext {
     }
 
     public static class MmtSignallingContext {
-        public int mpu_sequence_number;
+        public long mpu_sequence_number;
         public long mpu_presentation_time_ntp64;
         public long mpu_presentation_time_seconds;
         public long mpu_presentation_time_microseconds;
@@ -105,9 +105,9 @@ public class MmtPacketIdContext {
         }
 
         public Long last_computedPresentationTimestampUs = null;
-        public int extracted_sample_duration_us = 0;
+        public long extracted_sample_duration_us = 0;
 
-        public int last_mpu_sequence_number = 0;
+        public long last_mpu_sequence_number = 0;
         public int last_mfu_sample_number = 0;
         public Map<Integer, Long> last_mfu_release_microseconds = new HashMap<>();
